Move cover type
